Move sections/meetings to course instance
