+++ src/pages/courseInstance/courseInstance.js
@@ -1,0 +1,43 @@
+const router = require('express').Router({ mergeParams: true })
+
+const sections = [
+  {
+    name: 'Hello!',
+  },
+  {
+    name: 'World!',
+  },
+]
+
+const meetings = [
+  {
+    name: 'Goodbye!',
+  },
+  {
+    name: 'World!',
+  },
+]
+
+router.get('/', (req, res, next) => {
+  res.locals.courseInstanceId = req.params.courseInstanceId
+  res.locals.sections = sections
+  res.locals.meetings = meetings
+  res.render(__filename.replace(/\.js/, '.ejs'), res.locals)
+})
+
+router.post('/', (req, res, next) => {
+  if (req.body.__action === 'newSection') {
+    const section = {
+      name: req.body.name,
+    }
+    sections.push(section)
+  } else if (req.body.__action === 'newMeeting') {
+    const meeting = {
+      name: req.body.name,
+    }
+    meetings.push(meeting)
+  }
+  res.redirect(req.originalUrl)
+})
+
+module.exports = router
